Deleted unnecesary folder
